requirements.txt
Flask==2.3.3
gunicorn==21.2.0
Werkzeug==2.3.7
click==8.1.7
Jinja2==3.1.2
MarkupSafe==2.1.3
psycopg2-binary==2.9.7
SQLAlchemy==2.0.21
Flask-SQLAlchemy==3.0.5
python-dotenv==1.0.0 